requirements.txt
Django==1.11.4
djangorestframework==3.6.3
pytz==2017.2
South==1.0.2
